# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Music Platform Gateway API
  version: "2026-03-04"
  description: |
    Gateway API for the split architecture deployment.
    Base gateway: http://127.0.0.1:8080
    Verified with live tests on 2026-03-04.
servers:
  - url: http://127.0.0.1:8080
    description: Local gateway
  - url: http://slcdut.xyz:8080
    description: Public gateway
...
paths:
...
  /users/register:
    post:
      tags: [auth]
      summary: Register user
      operationId: registerUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            example:
              account: demo_account
              password: demo_password
              username: demo_user
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeSuccessBool"
              example:
                code: 0
                message: success
                data:
                  success: true
        "409":
          description: Account already exists
          content:
            text/plain:
              schema:
                type: string
              example: this account already exists
        "400":
          $ref: "#/components/responses/ErrorEnvelope"
  /users/login:
    post:
      tags: [auth]
      summary: Login user
      operationId: loginUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            example:
              account: demo_account
              password: demo_password
      responses:
        "200":
          description: Login response (success or fail)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
...
  /users/profile:
    get:
      tags: [auth]
      summary: Get current user profile
      operationId: getUserProfile
      parameters:
        - in: query
          name: account
          required: true
          schema:
            type: string
        - in: query
          name: session_token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeUserProfile"
        "401":
          $ref: "#/components/responses/ErrorEnvelope"
  /users/profile/username:
    post:
      tags: [auth]
      summary: Update current username
      operationId: updateUsername
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUsernameRequest"
            example:
              account: demo_account
              session_token: demo_session_token
              username: 新用户名
      responses:
        "200":
          description: Username updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeUpdateUsernameResponse"
        "401":
          $ref: "#/components/responses/ErrorEnvelope"
        "409":
          $ref: "#/components/responses/ErrorEnvelope"
  /users/profile/avatar:
    post:
      tags: [auth]
      summary: Upload or replace current avatar
      operationId: uploadAvatar
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [account, session_token, avatar]
              properties:
                account:
                  type: string
                session_token:
                  type: string
                avatar:
                  type: string
                  format: binary
      responses:
        "200":
          description: Avatar uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeUploadAvatarResponse"
        "401":
          $ref: "#/components/responses/ErrorEnvelope"
  /users/online/session/start:
    post:
      tags: [presence]
      summary: Start online session and issue session token
      operationId: startOnlineSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnlineSessionStartRequest"
      responses:
        "200":
          description: Session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeOnlineSession"
  /users/online/heartbeat:
    post:
      tags: [presence]
      summary: Heartbeat online session (requires session_token)
      operationId: heartbeatOnlineSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnlineHeartbeatRequest"
      responses:
        "200":
          description: Session renewed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeOnlineSession"
        "401":
          $ref: "#/components/responses/ErrorEnvelope"
  /users/online/status:
    get:
      tags: [presence]
      summary: Get current online status (requires session_token)
      operationId: getOnlineStatus
      parameters:
        - in: query
          name: account
          required: false
          schema:
            type: string
        - in: query
          name: username
          required: false
          schema:
            type: string
        - in: query
          name: session_token
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Online status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeOnlineStatus"
        "401":
          $ref: "#/components/responses/ErrorEnvelope"
  /users/online/logout:
    post:
      tags: [presence]
      summary: Logout online session (requires session_token)
      operationId: logoutOnlineSession
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OnlineLogoutRequest"
      responses:
        "200":
          description: Logout success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeSuccessBool"
        "401":
          $ref: "#/components/responses/ErrorEnvelope"
...
  /stream:
    post:
      tags: [catalog]
      summary: Get stream URL by filename
      description: Supports both local catalog filenames and Jamendo-compatible virtual filenames returned by `/music/search`.
      operationId: getMusicStreamByFilename
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StreamRequest"
            example:
              filename: hua_hai.mp3
      responses:
        "200":
          description: Music stream info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MusicResponse"
        "404":
          $ref: "#/components/responses/ErrorEnvelope"
  /get_music:
    get:
      tags: [catalog]
      summary: Get music detail by path
      operationId: getMusicByPath
      parameters:
        - in: query
          name: path
          required: true
          schema:
            type: string
          example: hua_hai/hua_hai.mp3
      responses:
        "200":
          description: Music detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MusicResponse"
        "404":
          $ref: "#/components/responses/ErrorEnvelope"
  /music/artist:
    post:
      tags: [catalog]
      summary: Search music by artist
      operationId: listMusicByArtist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtistRequest"
            example:
              artist: Jay Chou
      responses:
        "200":
          description: Music list filtered by artist
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileListItem"
  /music/search:
    post:
      tags: [catalog]
      summary: Search music by keyword
      description: |
        Uses keyword-based routing while keeping the legacy response schema.
        If the keyword contains Chinese characters, local catalog search is tried first and Jamendo is used as fallback.
        If the keyword contains Latin letters and no Chinese characters, Jamendo is tried first and local catalog search is used as fallback.
        Jamendo results are mapped into `FileListItem` with a virtual `path` that can be sent back to `/stream` and `/download`.
      operationId: searchMusic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
            example:
              keyword: hai
      responses:
        "200":
          description: Music search results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FileListItem"
  /external/music/jamendo/search:
    post:
      tags: [catalog]
      summary: Search Jamendo external music library
      description: Real-time search against Jamendo. Results are not stored locally and stream_url points to Jamendo audio.
      operationId: searchJamendoMusic
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExternalMusicSearchRequest"
            example:
              keyword: mayday
              limit: 20
      responses:
        "200":
          description: Jamendo external music results
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ExternalMusicTrack"
        "400":
          $ref: "#/components/responses/ErrorEnvelope"
        "502":
          $ref: "#/components/responses/ErrorEnvelope"
        "503":
          $ref: "#/components/responses/ErrorEnvelope"
  /external/music/jamendo/track:
    get:
      tags: [catalog]
      summary: Get Jamendo track detail
      description: Fetch one Jamendo track by Jamendo track id. This does not download or cache audio.
      operationId: getJamendoTrack
      parameters:
        - in: query
          name: id
          required: true
          schema:
            type: string
          example: "1218138"
      responses:
        "200":
          description: Jamendo external music detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ExternalMusicTrack"
        "400":
          $ref: "#/components/responses/ErrorEnvelope"
        "404":
          $ref: "#/components/responses/ErrorEnvelope"
        "502":
          $ref: "#/components/responses/ErrorEnvelope"
        "503":
          $ref: "#/components/responses/ErrorEnvelope"
  /artist/search:
    post:
      tags: [catalog]
      summary: Check if artist exists
      operationId: searchArtist
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtistRequest"
            example:
              artist: Jay Chou
      responses:
        "200":
          description: Artist existence
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistExistsResponse"
  /user/favorites/add:
    post:
      tags: [profile]
      summary: Add favorite music
      operationId: addFavorite
      description: Provide user identity with query param user_account or header X-User-Account.
      parameters:
        - in: query
          name: user_account
          required: false
          schema:
            type: string
        - in: header
          name: X-User-Account
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddFavoriteRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResponse"
        "400":
          description: Missing user account or invalid body
  /user/favorites:
    get:
      tags: [profile]
      summary: List favorites
      operationId: listFavorites
      description: Provide user identity with query param user_account or header X-User-Account.
      parameters:
        - in: query
          name: user_account
          required: false
          schema:
            type: string
        - in: header
          name: X-User-Account
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Favorite list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MusicItem"
        "400":
          description: Missing user account
  /user/favorites/remove:
    post:
      tags: [profile]
      summary: Remove favorite music
      operationId: removeFavorite
      description: user_account can be passed in query/header/body.
      parameters:
        - in: query
          name: user_account
          required: false
          schema:
            type: string
        - in: query
          name: music_path
          required: false
          schema:
            type: string
        - in: header
          name: X-User-Account
          required: false
          schema:
            type: string
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RemoveFavoriteRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResponse"
        "400":
          description: Missing parameters
  /user/history/add:
    post:
      tags: [profile]
      summary: Add play history record
      operationId: addPlayHistory
      description: Provide user identity with query param user_account or header X-User-Account.
      parameters:
        - in: query
          name: user_account
          required: false
          schema:
            type: string
        - in: header
          name: X-User-Account
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddPlayHistoryRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ActionResponse"
        "400":
          description: Missing user account or invalid body
  /user/history:
    get:
      tags: [profile]
      summary: List play history
      operationId: listPlayHistory
      description: Provide user identity with query param user_account or header X-User-Account.
      parameters:
        - in: query
          name: user_account
          required: false
          schema:
            type: string
        - in: header
          name: X-User-Account
          required: false
          schema:
            type: string
        - in: query
          name: limit
          required: false
          schema:
            type: integer
            default: 50
            minimum: 1
        - in: query
          name: distinct
          required: false
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: Play history
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MusicItem"
  /user/history/delete:
    post:
      tags: [profile]
      summary: Delete selected play history by music paths
      operationId: deletePlayHistory
      description: Provide user identity with query param user_account or header X-User-Account.
      parameters:
        - in: query
          name: user_account
          required: false
          schema:
            type: string
        - in: header
          name: X-User-Account
          required: false
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeletePlayHistoryRequest"
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteActionResponse"
  /user/history/clear:
    post:
      tags: [profile]
      summary: Clear all play history
      operationId: clearPlayHistory
      description: Provide user identity with query param user_account or header X-User-Account.
      parameters:
        - in: query
          name: user_account
          required: false
          schema:
            type: string
        - in: header
          name: X-User-Account
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Success
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteActionResponse"
...
  /admin/recommend/retrain:
    post:
      tags: [recommendation]
      summary: Trigger recommendation retrain task
      operationId: triggerRecommendationRetrain
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RecommendationRetrainRequest"
      responses:
        "202":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EnvelopeTaskAccepted"
...
components:
  responses:
    ErrorEnvelope:
      description: Error response envelope
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorEnvelope"
  schemas:
...
    EnvelopeSuccessBool:
      type: object
      properties:
        code:
          type: integer
          example: 0
        message:
          type: string
          example: success
        data:
          type: object
          properties:
            success:
              type: boolean
          required: [success]
      required: [code, message, data]
...
    RegisterRequest:
      type: object
      required: [account, password, username]
      properties:
        account:
          type: string
        password:
          type: string
        username:
          type: string
    LoginRequest:
      type: object
      required: [account, password]
      properties:
        account:
          type: string
        password:
          type: string
    LoginResponse:
      type: object
      properties:
        success:
          type: string
          enum: ["true", "false"]
        success_bool:
          type: boolean
        username:
          type: string
        avatar_url:
          type: string
        song_path_list:
          type: array
          items:
            type: string
        online_session_token:
          type: string
        online_heartbeat_interval_sec:
          type: integer
        online_ttl_sec:
          type: integer
      required: [success, success_bool, username, song_path_list]
...
    OnlineSessionStartRequest:
      type: object
      properties:
        account:
          type: string
        username:
          type: string
        device_id:
          type: string
...